add logger log4j1 to console remove UserProfile form User add lombok
