+++ src/main/java/client/model/formatMsgWithServer/MessageFromServer.java
@@ -2,6 +2,7 @@ package client.model.formatMsgWithServer;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+
 import java.sql.Timestamp;
 
 public class MessageFromServer {
