+++ src/main/java/client/view/RootLayoutController.java
@@ -3,6 +3,7 @@ package client.view;
 import client.controller.ClientController;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+
 import java.net.URL;
 import java.util.ResourceBundle;
