+++ src/main/java/database/ExampleUsage.java
@@ -1,11 +1,8 @@
 package database;
 
 import database.dao.DataBaseService;
-import database.entity.Message;
 import database.entity.User;
 
-import java.sql.Time;
-import java.sql.Timestamp;
 import java.util.List;
 
 /**
