+++ src/main/java/database/entity/Message.java
@@ -2,7 +2,6 @@ package database.entity;
 
 
 import javax.persistence.*;
-import java.sql.Time;
 import java.sql.Timestamp;
 
 @Entity
